remove unneeded copy buffer func
