+++ app/src/test/java/com/scepticalphysiologist/dmaple/Scrap.kt
@@ -1,14 +1,12 @@
 package com.scepticalphysiologist.dmaple
 
-import com.scepticalphysiologist.dmaple.map.buffer.copyBuffer
-import mil.nga.tiff.TIFFImage
+
 import mil.nga.tiff.TiffReader
 import org.junit.Test
 import java.io.File
 import java.io.RandomAccessFile
 import java.nio.ByteBuffer
 import java.nio.channels.FileChannel
-import kotlin.math.floor
 import kotlin.random.Random
 import kotlin.system.measureTimeMillis
 
@@ -113,8 +111,28 @@ class Scrap {
 
     }
 
-
-
-
-
 }
+
+
+/** Copy the content of one buffer to another. */
+fun copyBuffer(src: ByteBuffer, dst: ByteBuffer) {
+    // Make sure we are copying from the start of both buffers.
+    src.position(0)
+    dst.position(0)
+
+    // Adjust the limit of the source to that of the destination,
+    // so we don't get buffer overflow on the destination.
+    val srcOldLimit = src.limit()
+    if(dst.limit() < src.limit()) src.limit(dst.limit())
+
+    // Do the copy, setting the destination endianness to that of the source.
+    try {
+        dst.put(src)
+        dst.order(src.order())
+    } catch(_: java.nio.BufferOverflowException) {}
+
+    // Re-set the source limit.
+    src.limit(srcOldLimit)
+}
+
+
